requirements.txt
# Main dependencies
fastapi
langchain-milvus
pymilvus >= 2.5.5
uvicorn
python-dotenv
pydantic
pydantic-settings
openai>=1.30.0
langchain
langchain-openai
langchain-community
requests
tenacity
sqlmodel
pymysql
cryptography
sentence-transformers
httpx
pytest

# OpenAI Agents
openai-agents>=0.2.5

# Utility dependencies
python-multipart

